Remove 'else' keyword, in favour of if/if not syntax.
I prefer the Plan 9-esque if/if not syntax to the if..else introduced in the
rakitzis Unix port.

This commit removes the 'else' token from the parser and tree-walk interpreter,
and refactors the latter to correct the logic for handling 'if not's following
'if's, and the common 'if not if...' idiom.