Introduce a 'noexport' variable, from es(1).
Some variables are never exported to the environment, most of which are related
to the internal functioning of rc(1). This list was previously harcoded,
however later shells in the rc family (es, xs) included a 'noexport' variable,
which contained the name of variables which should never be exported to the
environment. This commit ports this 'noexport' feature to rc.