Improve readline string quoting function.
The readline interactive editing support code includes a function which quotes
special characters in an input buffer, which is used during tab completion for
handling e.g. filenames with spaces in them.

The previous version of this code would use the mprint() machinery to simply
wrap the input word in quotes (and handle any single quote characters within
the input word), which doesn't work particularly well with words which contain
slashes in them (e.g. completing into a subdirectory).

This commit reimplements the quoting function so that only runs of special
chars in the input word are quoted, and other regular chars are left
alone. This (ab)uses the parser's insertion of free carets for concatenating
the resulting alternating quote/non-quoted word fragments into a single word.

(Tab completion of variable names which include special characters which need
to be quoted is broken further by this commit, but it didn't really work in the
first place.)